Added DAISY_PROFILING optional code parts and DAISY_HOST_TRANSFERS parts to program, making transfers back to cpu an optional function.
Also the program can now sort of handle non-512x512 multiples of image sizes,
but not non-16 multiples of block heights, this can and will be easily
fixed in the next commit.

Also, the single-block case has been sped up by 20-30ms by not copying
the data from pinned memory to pageable on the host. The data seems to be
usable normally from pinned memory.

todo: - fix the irregular dimension handling in the blocks
      - fix the memory problems of large images on GPU, a limit will have to be introduced by the program
          unless large image handling is implemented (the thought was 1024x1024) but not now.

+++ src/daisy/daisyKernels.cl
@@ -80,11 +80,11 @@ __kernel void gradient_8all(__global float * massArray,
   const int c = get_global_id(0) % pddWidth;
   const int srcOffset = pddWidth * pddHeight * 8 + r * pddWidth + c;
 
-  short4 n;
-  n.x = (short)(c > 0           ? massArray[srcOffset-1]:massArray[srcOffset]);
-  n.y = (short)(r > 0           ? massArray[srcOffset-pddWidth]:massArray[srcOffset]);
-  n.z = (short)(c < pddWidth-1  ? massArray[srcOffset+1]:massArray[srcOffset]);
-  n.w = (short)(r < pddHeight-1 ? massArray[srcOffset+pddWidth]:massArray[srcOffset]);
+  float4 n;
+  n.x = (c > 0           ? massArray[srcOffset-1]:massArray[srcOffset]);
+  n.y = (r > 0           ? massArray[srcOffset-pddWidth]:massArray[srcOffset]);
+  n.z = (c < pddWidth-1  ? massArray[srcOffset+1]:massArray[srcOffset]);
+  n.w = (r < pddHeight-1 ? massArray[srcOffset+pddWidth]:massArray[srcOffset]);
 
   float8 gradients;
   const float8 angles = (float8)(0.0f, M_PI / 4, M_PI / 2, 3 * (M_PI / 4), M_PI,
